# backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: File Share API
  description: REST API for file listing, upload, download, and directory management.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
  - url: http://localhost:22010
    description: Docker / production (HTTP behind reverse proxy)
security:
  - basicAuth: []
paths:
  /api/files:
    get:
      summary: List directory contents
      description: Returns the shared page envelope; pass nextCursor back verbatim to continue.
      parameters:
        - name: path
          in: query
          schema:
            type: string
            example: ""
        - name: limit
          in: query
          description: Page size (default 50, clamped to 1..500)
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: cursor
          in: query
          description: Opaque next cursor from a previous page
          schema:
            type: string
      responses:
        "200":
          description: One page of file entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilePage"
        "400":
          description: Invalid path or malformed cursor
        "401":
          description: Unauthorized
        "404":
          description: Not a directory
    delete:
      summary: Delete a file or directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
      responses:
        "200":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /api/files/download:
    get:
      summary: Download a file or ZIP archive
      description: >
        Regular files carry a strong ETag (name+size+mtime) and support byte
        ranges: If-None-Match revalidates to 304, Range/If-Range serve 206
        partial content (416 when unsatisfiable). ZIP archives are synthetic
        streams and are always sent whole with a 200.
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: Range
          in: header
          description: Byte range, e.g. bytes=0-1023 (seekable files only)
          schema:
            type: string
        - name: If-None-Match
          in: header
          description: ETag from a previous response; matches answer 304
          schema:
            type: string
      responses:
        "200":
          description: File or ZIP stream
        "206":
          description: Partial content matching the requested byte range
        "304":
          description: Not modified (validator matched)
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "416":
          description: Requested range not satisfiable
  /api/files/info:
    get:
      summary: Get file or directory metadata
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
        "404":
          description: Not found
  /api/files/search:
    get:
      summary: Search files by name, path, or full-text content
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            example: report
        - name: limit
          in: query
          description: Page size (default 50, clamped to 1..500)
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: cursor
          in: query
          description: Opaque next cursor from a previous page
          schema:
            type: string
      responses:
        "200":
          description: One page of matches (name/path index plus full-text content index)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilePage"
        "400":
          description: Missing or invalid query, or malformed cursor
  /api/events:
    get:
      summary: Live filesystem events (Server-Sent Events)
      description: |
        Long-lived SSE stream of upload/update/mkdir/delete/share/restore/quota
        events for the authenticated user. Browser clients use the HttpOnly
        session cookie; API clients may send Authorization: Bearer.
      responses:
        "200":
          description: text/event-stream of JSON event frames
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"type":"upload","path":"a.txt","user":"alice","at":"2026-01-01T00:00:00Z"}
        "401":
          description: Unauthorized
        "405":
          description: Method not allowed
  /api/p2p/stream:
    get:
      summary: WebRTC signaling stream (Server-Sent Events)
      description: |
        Registers this browser tab as a peer session. Frames: hello (own
        peerId + current peers), peers (presence updates), signal (incoming
        offer/answer/candidate/bye). File bytes never pass through this API.
      responses:
        "200":
          description: text/event-stream of signaling frames
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: hello
                  data: {"type":"hello","peerId":"a1b2","peers":[],"at":"2026-01-01T00:00:00Z"}
        "401":
          description: Unauthorized
        "405":
          description: Method not allowed
  /api/p2p/peers:
    get:
      summary: List connected P2P peer sessions
      responses:
        "200":
          description: Connected peers
          content:
            application/json:
              schema:
                type: object
                properties:
                  peers:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        user:
                          type: string
        "401":
          description: Unauthorized
  /api/p2p/signal:
    post:
      summary: Relay one WebRTC signal to a peer
      parameters:
        - in: header
          name: X-Peer-Id
          required: true
          schema:
            type: string
          description: Own peer id from the hello frame on /api/p2p/stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, kind]
              properties:
                to:
                  type: string
                  description: Destination peer id
                kind:
                  type: string
                  enum: [offer, answer, candidate, bye]
                payload:
                  type: object
                  additionalProperties: true
                  description: Opaque SDP or ICE candidate JSON
      responses:
        "202":
          description: Signal queued for delivery
        "400":
          description: Invalid body or signal
        "403":
          description: X-Peer-Id does not match a live session for the caller
        "404":
          description: Destination peer offline
        "401":
          description: Unauthorized
  /api/upload:
    post:
      summary: Upload one or more files
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                path:
                  type: string
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "413":
          description: Upload exceeds MAX_UPLOAD_BYTES
        "401":
          description: Unauthorized
  /api/uploads:
    get:
      summary: List active resumable upload sessions for the caller
      description: |
        Pending / interrupted uploads (offset < size or ready to complete)
        that can be resumed after a reload. Sessions are scoped to the
        authenticated owner (or the system view when auth is disabled).
      responses:
        "200":
          description: Active sessions, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    fingerprint:
                      type: string
                    destination:
                      type: string
                    filename:
                      type: string
                    size:
                      type: integer
                      format: int64
                    offset:
                      type: integer
                      format: int64
                    updatedAt:
                      type: string
                      format: date-time
                    expiresAt:
                      type: string
                      format: date-time
        "401":
          description: Unauthorized
    post:
      summary: Create (or resume) a resumable upload session
      description: |
        Returns the session id, current staged offset, and preferred chunkSize.
        When fingerprint matches an existing active session for the caller,
        that session is returned so the client can continue from offset.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, size]
              properties:
                path:
                  type: string
                  description: Destination directory (virtual path)
                filename:
                  type: string
                size:
                  type: integer
                  format: int64
                fingerprint:
                  type: string
                  description: Client identity for resume (path|name|size|mtime)
      responses:
        "200":
          description: Session created or resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadSession"
        "401":
          description: Unauthorized
        "413":
          description: Size exceeds quota or MAX_UPLOAD_BYTES
...
  /api/directories:
    post:
      summary: Create a directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
      responses:
        "201":
          description: Directory created
        "401":
          description: Unauthorized
  /api/auth/password:
    post:
      summary: Change the authenticated user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        "200":
          description: Password changed
        "400":
          description: Validation failed
        "401":
          description: Invalid credentials
  /api/auth/api-keys:
    get:
      summary: List the caller's API keys (metadata only)
      responses:
        "200":
          description: Key list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/APIKey"
        "401":
          description: Unauthorized
        "403":
          description: Authentication required
    post:
      summary: Mint an API key (plaintext returned once)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 80
                scope:
                  type: string
                  enum: [read, write, admin]
                  default: read
                expiresIn:
                  type: integer
                  format: int64
                  description: Seconds until expiry (0 or absent = never)
      responses:
        "201":
          description: Key created; `key` is shown exactly once
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    example: sfs_0123456789abcdef_AbCdEfGhIjKlMnOpQrStUvWxYz01234567
                  id:
                    type: string
                  name:
                    type: string
                  scope:
                    type: string
                  prefix:
                    type: string
        "400":
          description: Validation failed
        "401":
          description: Unauthorized
        "403":
          description: Authentication required
  /api/auth/api-keys/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      summary: Revoke one of the caller's API keys
      responses:
        "200":
          description: Key revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [revoked]
        "401":
          description: Unauthorized
        "403":
          description: Authentication required
        "404":
          description: Unknown key id
  /api/admin/users:
    get:
      summary: List accounts with storage usage (users.read)
      parameters:
        - name: limit
          in: query
          description: Page size (default 50, clamped to 1..500)
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: cursor
          in: query
          description: Opaque next cursor from a previous page
          schema:
            type: string
      responses:
        "200":
          description: One page of user stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStatsPage"
        "400":
          description: Malformed cursor
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    post:
      summary: Create an account (users.create)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  example: member
                enabled:
                  type: boolean
                  default: true
                quotaBytes:
                  type: integer
                  format: int64
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation failed
        "409":
          description: Username already exists
        "403":
          description: Forbidden
...
  /api/admin/users/{username}/quota:
    parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
    put:
      summary: Set an account storage quota (quota.manage)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quota]
              properties:
                quota:
                  type: string
                  example: 2GB
      responses:
        "200":
          description: Updated user stats
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    FileItem:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        size:
          type: integer
          format: int64
        isDir:
          type: boolean
        modified:
          type: string
          format: date-time
...
    UploadResult:
      type: object
      properties:
        path:
          type: string
        size:
          type: integer
          format: int64
    UploadSession:
      type: object
      properties:
        id:
          type: string
        offset:
          type: integer
          format: int64
          description: Bytes already staged
        size:
          type: integer
          format: int64
          description: Total declared size
        chunkSize:
          type: integer
          format: int64
          description: Preferred chunk size for PATCH bodies
        expiresAt:
          type: string
          format: date-time
...
    FilePage:
      type: object
      description: Shared list envelope; nextCursor is absent on the last page
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FileItem"
        nextCursor:
          type: string
          description: Opaque cursor for the next page; absent on the last page
...